Add lane recognition to camera
The camera can recognize dark spots in the image. 
The code checks different horizontal lines in the image for dark spots (going from the middle to the border). If the pixel is under a threshold for a certain amount of pixels it stores the value and draws a circle over it.