Implement virtual_cam.py
Implements virtual_cam.py, which is used to process videos captured by the OpenMV cam without the need to have the camera at hand.
It uses a video defined by set_input_and_output (paths have to change based on the computer) and applies the lane recognition and steering functions, which are also drastically improved compared to old versions.
At last debug visuals (such as the speed and steering) can be drawn on to the video.

As the camera is coded in Python as well, porting the shared code should be easy.