Fix Issue #10
Adds functions inside __init__.py of both lane recognition and movement params. These functions return the id of a lane recognition algorithm.
This removes the need to update the functions in main.py and virtual_cam.py when you add a new algorithm.

Update the Create New Instance.md to reflect these changes and what you have to do in the __init__.py