Add fault handling
If the camera crashes it will now create a log with the error that occurred. Additionally, it should not crash because of that anymore (It just logs the error and continues).
If the camera somehow still crashes it will restart after one second.

The i2c bus is now restarted if now connection is possible (this did not yet have an effect on the i2c bus stability).