Add Communication Management
Adds a dedicated package for communication management.

This makes it easier to implement different communication solutions.

Currently, the communication with the Teensy is implemented via i2c. The SPI from the Samples folder was implemented as well.

To change the communication handler you simply have to enter a different name for the communication handler when setting it up.

NOT TESTED ON REAL HARDWARE.